move record unification to own file
